Add files via upload
Changed to manually add Spaces to make companies appear as aligned columns after receiving an email. Can fill in the form and send the email to the company recipient email. Use html to auto-align code in comments. The page is written in html and JS, and the email function is written in PHP.